Add mocking for fused location provider
Fused location provider can occasionally be slow to update. This means for a short moment, the real location of the phone can be returned to apps rather than the mocked location. If the user is using this app to hide his real location, this can be undesirable. This PR fixes this issue by mocking the fused location provider as well.